Revert "file centric Cart refactor. (Body, Header, DialogBox)"
