--- src/pages/fileCentricCart/components/body/cartBody.style.js
@@ -1,3 +1,0 @@
-export default () => ({
-
-});
