--- src/pages/fileCentricCart/components/dialogBox/dialogBox.style.js
@@ -1,18 +1,0 @@
-export default () => ({
-  okButton: {
-    background: '#98A19E',
-    color: '#fff',
-    cursor: 'pointer',
-    '&:hover': {
-      backgroundColor: 'rgba(152,161,158,0.6)',
-    },
-  },
-  cancelButton: {
-    background: '#42779A',
-    color: '#fff',
-    cursor: 'pointer',
-    '&:hover': {
-      backgroundColor: 'rgba(66,119,154,0.6)',
-    },
-  },
-});
