--- src/pages/fileCentricCart/components/footer/cartFooter.style.js
@@ -1,3 +1,0 @@
-export default () => ({
-
-});
