--- src/pages/fileCentricCart/components/header/cartHeader.style.js
@@ -1,35 +1,0 @@
-export default (theme) => ({
-  logo: {
-    position: 'absolute',
-    float: 'left',
-    marginTop: '46px',
-    width: '100px',
-  },
-  header: {
-    paddingLeft: '32px',
-    paddingRight: '32px',
-    borderBottom: '#81a6b9 4px solid',
-    height: '140px',
-    margin: 'auto 33px',
-  },
-  headerTitle: {
-    maxWidth: theme.custom.maxContentWidth,
-    margin: 'auto',
-    float: 'left',
-    marginLeft: '110px',
-    paddingLeft: '3px',
-  },
-  headerMainTitle: {
-    fontFamily: theme.custom.fontFamilySans,
-    fontWeight: 'bold',
-    letterSpacing: '0.017em',
-    color: '#ff8a00',
-    fontSize: '30px',
-    lineHeight: '18px',
-    paddingLeft: '5px',
-    paddingBottom: '8px',
-  },
-  marginTop80: {
-    marginTop: '80px',
-  },
-});
